Restore some of my earlier edits and formatting.
Transition to a switch() statement for MSP receiver state machine,
since this leads to better code if the case statements are testing
a contiguous range of integers.  And it is far more readable.
Also use an accessor for changing the state in case we need
diagnostics in there or a trap under a debugger.