Refactored to flat (no module) structure.
Also tweaked the build system. It's probably hopelessly broken on
Windows/Lunix for now, but it runs on OS X.

Changed rxtx dependency to rxtx-rebundled (to avoid jni packaging
drama). I'm not 100% convinced that adding a foreign dep is the best way
to go, but it gets us up and running, and we can always re-evaluate.